refactor: change project structure, change implementation of set_joints function in head
