+++ CMakeLists.txt
@@ -26,9 +26,8 @@ find_package(tachimawari REQUIRED)
 find_package(tachimawari_interfaces REQUIRED)
 
 add_library(${PROJECT_NAME} SHARED
-  "src/${PROJECT_NAME}/head/head.cpp"
-  "src/${PROJECT_NAME}/receiver/node/receiver_node.cpp"
-  "src/${PROJECT_NAME}/sender/node/sender_node.cpp"
+  "src/${PROJECT_NAME}/head/process/head.cpp"
+  "src/${PROJECT_NAME}/head/node/head_node.cpp"
   "src/${PROJECT_NAME}/node/atama_node.cpp")
 
 target_include_directories(${PROJECT_NAME} PUBLIC
@@ -55,13 +54,13 @@ install(TARGETS ${PROJECT_NAME}
   LIBRARY DESTINATION "lib"
   RUNTIME DESTINATION "bin")
 
-add_executable(atama_main "src/atama_main.cpp")
-target_include_directories(atama_main PUBLIC
+add_executable(main "src/atama_main.cpp")
+target_include_directories(main PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
-target_link_libraries(atama_main ${PROJECT_NAME})
+target_link_libraries(main ${PROJECT_NAME})
 
-install(TARGETS atama_main
+install(TARGETS main
   DESTINATION lib/${PROJECT_NAME})
 
 if(BUILD_TESTING)
